Version .002
Changed the name to "tcldev.vim" to better reflect its intent!

Added folding

Added some widget stuff back in (not tested!)

Changed the "-" option to make sure it only colors alphas and not numerics

Lots of this was done by Tomasz Kalkosiński!!

The intent is to get this up to speed and then submit it to the current maintainer of "tcl.vim".

TODO: Tomasz implemented a tclOperator section that isn't in working condition yet.

TODO: Color the following like /* */

if {0} {

}

TODO: Whatever else floats our boat?  : )